third_party/blink/web_tests/external/wpt/html/editing/the-hidden-attribute/hidden-until-found-and-details-expected.txt

This is a testharness.js-based test.
[FAIL] hidden=until-found and details revealing algorithm should reveal elements in a certain order.
  assert_equals: a3 should not be open expected false but got true
Harness: the test ran to completion.
